feat: added upload, update and delete images functionality to experience form
refactor: update useMediaUploader custom hook, experience services, cloudinary services

+++ backend/services/experience.service.js
@@ -1,4 +1,8 @@
 import Experience from "../model/experience.model.js";
+import {
+  updateCloudinaryMedia,
+  deleteCloudinaryMedia,
+} from "./cloudinary/cloudinary.service.js";
 
 const createExperienceService = async (data) => {
   try {
@@ -13,6 +17,22 @@ const createExperienceService = async (data) => {
 
 const updateExperienceService = async (id, data) => {
   try {
+    const experience = await Experience.findById(id);
+    if (!experience) {
+      throw new Error("Project not found");
+    }
+
+    await updateCloudinaryMedia(
+      {
+        logoPublicId: experience.logoPublicId,
+        iconPublicId: experience.iconPublicId,
+      },
+      {
+        logoPublicId: data.logoPublicId,
+        iconPublicId: data.iconPublicId,
+      }
+    );
+
     const updatedExperience = await Experience.findByIdAndUpdate(id, data, {
       new: true,
     });
@@ -28,14 +48,25 @@ const updateExperienceService = async (id, data) => {
 
 const deleteExperienceService = async (id) => {
   try {
+    const experience = await Experience.findById(id);
+    if (!experience) {
+      throw new Error("Experience not found");
+    }
+
+    await deleteCloudinaryMedia({
+      logoPublicId: experience.logoPublicId,
+      iconPublicId: experience.iconPublicId,
+    });
+
     const deletedExperience = await Experience.findByIdAndDelete(id);
     if (!deletedExperience) {
-      throw new Error("Employment history not found");
+      throw new Error("Experience deletion failed");
     }
+
     return deletedExperience;
   } catch (error) {
     console.error(error);
-    throw new Error("Failed to delete employment history. Please try again.");
+    throw new Error("Failed to delete experience. Please try again.");
   }
 };
